General(:zap:): Abstracting pullSwitch animation.
